remove PatientSearchbar
