+++ src/components/Inputs/index.jsx
@@ -2,15 +2,7 @@ import InputNumber from './InputNumber/InputNumber'
 import InputDate from './InputDate/InputDate'
 import InputAutocompleteAsync from './InputAutocompleteAsync/InputAutocompleteAsync'
 
-import InputSearchDocumentExtend from './InputSearchDocument/components/InputSearchDocumentExtend'
 import InputSearchDocumentSimple from './InputSearchDocument/components/InputSearchDocumentSimple'
 import InputAgeRange from './InputAgeRange/InputAgeRange'
 
-export {
-  InputNumber,
-  InputDate,
-  InputAutocompleteAsync,
-  InputSearchDocumentExtend,
-  InputSearchDocumentSimple,
-  InputAgeRange
-}
+export { InputNumber, InputDate, InputAutocompleteAsync, InputSearchDocumentSimple, InputAgeRange }
